backend/requirements.txt
# FastAPI and ASGI server (updated for anyio compatibility)
fastapi>=0.115.0
uvicorn[standard]>=0.25.0

# Database
sqlalchemy>=2.0.23
alembic>=1.12.1

# Pydantic for data validation
pydantic>=2.9.2
pydantic-settings>=2.1.0

# Browser automation with latest browser-use
playwright>=1.40.0
browser-use>=0.5.4
patchright>=1.0.0

# HTTP client (for making requests)
httpx>=0.25.2

# Environment variables
python-dotenv>=1.0.1

# Logging and utilities
python-json-logger==2.0.7

# Development dependencies
pytest>=8.3.3
pytest-asyncio>=0.24.0
black==23.11.0
isort==5.12.0 